Question, Comment 분리
