+++ src/main/kotlin/hntech/hntechserver/archive/ArchiveController.kt
@@ -1,6 +1,6 @@
 package hntech.hntechserver.archive
 
-import Auth
+import hntech.hntechserver.utils.auth.Auth
 import hntech.hntechserver.utils.BoolResponse
 import hntech.hntechserver.utils.config.PAGE_SIZE
 import hntech.hntechserver.utils.logger
@@ -33,17 +33,20 @@ class ArchiveController(
      * 관리자 모드
      */
     // 자료글 생성
-    @Auth @PostMapping
+    @Auth
+    @PostMapping
     fun createArchive(@RequestBody form: ArchiveForm) =
         ArchiveSimpleResponse(archiveService.createArchive(form))
 
     // 자료글 수정
-    @Auth @PutMapping("/{archiveId}")
+    @Auth
+    @PutMapping("/{archiveId}")
     fun updateArchive(@PathVariable("archiveId") id: Long, @RequestBody form: ArchiveForm) =
         ArchiveSimpleResponse(archiveService.updateArchive(id, form))
 
     // 자료글 삭제
-    @Auth @DeleteMapping("/{archiveId}")
+    @Auth
+    @DeleteMapping("/{archiveId}")
     fun deleteArchive(@PathVariable("archiveId") id: Long) =
         BoolResponse(archiveService.deleteArchive(id))
