+++ src/main/kotlin/hntech/hntechserver/comment/CommentRepository.kt
@@ -1,0 +1,7 @@
+package hntech.hntechserver.comment
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CommentRepository: JpaRepository<Comment, Long> {
+    fun findAllByQuestionId(questionId: Long): List<Comment>
+}
