+++ src/main/kotlin/hntech/hntechserver/utils/auth/Auth.kt
@@ -1,3 +1,5 @@
+package hntech.hntechserver.utils.auth
+
 @Target(AnnotationTarget.FUNCTION)
-@kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
+@Retention(AnnotationRetention.RUNTIME)
 annotation class Auth
