+++ src/test/kotlin/hntech/hntechserver/comment/CommentServiceTest.kt
@@ -1,13 +1,16 @@
-package hntech.hntechserver.question
+package hntech.hntechserver.comment
 
-import hntech.hntechserver.question.dto.*
+import hntech.hntechserver.question.*
+import hntech.hntechserver.question.dto.CommentCreateForm
+import hntech.hntechserver.question.dto.CommentResponse
+import hntech.hntechserver.question.dto.CommentUpdateForm
+import hntech.hntechserver.question.dto.QuestionCreateForm
 import hntech.hntechserver.utils.logger
-import org.assertj.core.api.Assertions.*
-import org.junit.jupiter.api.Test
-
-import org.junit.jupiter.api.Assertions.*
+import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.transaction.annotation.Transactional
